EZP-29108: Add page factory
