+++ features/ContentCreation.feature
@@ -14,8 +14,8 @@ Scenario: Content creation can be cancelled
       | label | value        |
       | Title | Test Article |
     And I click on the edit action bar button "Cancel"
-  Then I should be on content container page "eZ Platform" of type "Folder"
-    And there's no "Test Article" "Article" on "eZ Platform" Sub-items list
+  Then I should be on root container page in Content View
+    And there's no "Test Article" "Article" on Sub-items list of root
 
 @javascript @common
 Scenario: Content creation can be closed
@@ -24,8 +24,8 @@ Scenario: Content creation can be closed
       | label | value        |
       | Title | Test Article |
     And I click on the close button
-  Then I should be on content container page "eZ Platform" of type "Folder"
-    And there's no "Test Article" "Article" on "eZ Platform" Sub-items list
+  Then I should be on root container page in Content View
+    And there's no "Test Article" "Article" on Sub-items list of root
 	
 @javascript @common
 Scenario: Content can be previewed during creation
@@ -62,7 +62,7 @@ Scenario: Content draft can be published
   Given going to dashboard we see there's draft "Test Article2" on list
   When I start editing content draft "Test Article2"
     And I click on the edit action bar button "Publish"
-  Then I should be on content container page "Test Article2" of type "Article" in "eZ Platform"
+  Then I should be on content container page "Test Article2" of type "Article" in root path
     And going to dashboard we see there's no draft "Test Article2" on list
 
 @javascript @common
@@ -74,7 +74,7 @@ Scenario: Content draft can be deleted
       | Title | Test Article edited       |
       | Intro | Test article intro edited |
     And I click on the edit action bar button "Delete draft"
-  Then I should be on content container page "eZ Platform" of type "Folder"
+  Then I should be on root container page in Content View
     And going to dashboard we see there's no draft "Test Article edited" on list
 
 @javascript @common
@@ -85,7 +85,7 @@ Scenario: Content can be published
       | Title | Test Article       |
       | Intro | Test article intro |
     And I click on the edit action bar button "Publish"
-  Then I should be on content container page "Test Article" of type "Article" in "eZ Platform"
+  Then I should be on content container page "Test Article" of type "Article" in root path
     And success notification that "Content published." appears
     And content attributes equal
       | label | value              |
@@ -101,7 +101,7 @@ Scenario: Content edit draft can be deleted
       | Title | Test Article edited       |
       | Intro | Test article intro edited |
     And I click on the edit action bar button "Delete draft"
-  Then I should be on content container page "Test Article" of type "Article" in "eZ Platform"
+  Then I should be on content container page "Test Article" of type "Article" in root path
     And going to dashboard we see there's no draft "Test Article edited" on list
 
 @javascript @common
@@ -127,7 +127,7 @@ Scenario: Content draft edition can be closed
       | Title | Test Article edited2       |
       | Intro | Test article intro edited2 |
     And I click on the close button
-  Then I should be on content container page "Test Article" of type "Article" in "eZ Platform"
+  Then I should be on content container page "Test Article" of type "Article" in root path
 
 @javascript @common
 Scenario: Content edit draft can be saved
@@ -151,7 +151,7 @@ Scenario: Content draft can be created and published through draft list modal
       | label | value                      |
       | Title | Test Article edited3       |
     And I click on the edit action bar button "Publish"
-  Then I should be on content container page "Test Article edited3" of type "Article" in "eZ Platform"
+  Then I should be on content container page "Test Article edited3" of type "Article" in root path
     And success notification that "Content published." appears
     And content attributes equal
       | label | value                |
@@ -182,7 +182,7 @@ Scenario: Content draft from draft list modal can be published
       | label | value                      |
       | Title | Test Article edited4       |
     And I click on the edit action bar button "Publish"
-  Then I should be on content container page "Test Article edited4" of type "Article" in "eZ Platform"
+  Then I should be on content container page "Test Article edited4" of type "Article" in root path
     And success notification that "Content published." appears
     And content attributes equal
       | label | value                |
@@ -195,12 +195,12 @@ Scenario: Content moving can be cancelled
   When I click on the edit action bar button "Move"
     And I select content "Media/Images" through UDW
     And I close the UDW window
-  Then I should be on content container page "Test Article edited4" of type "Article" in "eZ Platform"
+  Then I should be on content container page "Test Article edited4" of type "Article" in root path
     And content attributes equal
       | label | value                |
       | Title | Test Article edited4 |
       | Intro | Test article intro   |
-    And breadcrumb shows "eZ Platform/Test Article edited4" path
+    And breadcrumb shows "Test Article edited4" path under root path
 
 @javascript @common
 Scenario: Content can be moved
@@ -221,7 +221,7 @@ Scenario: Content can be moved
 Scenario: Content copying can be cancelled
   Given I navigate to content "Test Article edited4" of type "Article" in "Media/Images"
   When I click on the edit action bar button "Copy"
-    And I select content "eZ Platform" through UDW
+    And I select content root node through UDW
     And I close the UDW window
   Then I should be on content container page "Test Article edited4" of type "Article" in "Media/Images"
     And content attributes equal
@@ -234,10 +234,10 @@ Scenario: Content copying can be cancelled
 Scenario: Content can be copied
   Given I navigate to content "Test Article edited4" of type "Article" in "Media/Images"
   When I click on the edit action bar button "Copy"
-    And I select content "eZ Platform" through UDW
+    And I select content root node through UDW
     And I confirm the selection in UDW
-  Then I should be on content container page "Test Article edited4" of type "Article" in "eZ Platform"
-    And success notification that "'Test Article edited4' copied to 'eZ Platform'" appears
+  Then I should be on content container page "Test Article edited4" of type "Article" in root path
+    And success notification that "Test Article edited4" has been copied to root node appears
     And content attributes equal
       | label | value                |
       | Title | Test Article edited4 |
