+++ src/lib/Behat/Helper/UtilityContext.php
@@ -6,11 +6,12 @@
  */
 namespace EzSystems\EzPlatformAdminUi\Behat\Helper;
 
-use Behat\Mink\Element\NodeElement;
-use Behat\Mink\Element\TraversableElement;
+use Behat\Mink\Element\ElementInterface;
 use Behat\Mink\Exception\ElementNotFoundException;
 use Behat\MinkExtension\Context\MinkContext;
 use Exception;
+use EzSystems\EzPlatformAdminUi\Behat\Helper\NullElement\NullElement;
+use EzSystems\EzPlatformAdminUi\Behat\Helper\NullElement\NullElementException;
 use WebDriver\Exception\ElementNotVisible;
 
 class UtilityContext extends MinkContext
@@ -41,11 +42,11 @@ class UtilityContext extends MinkContext
      * to find all element with a given css selector that might still be loading.
      *
      * @param   string      $locator        css selector for the element
-     * @param   TraversableElement $baseElement    base Mink node element from where the find should be called
+     * @param   ElementInterface $baseElement    base Mink node element from where the find should be called
      *
-     * @return  NodeElement[]
+     * @return  ElementInterface[]
      */
-    public function findAllElements(string $locator, TraversableElement $baseElement = null): array
+    public function findAllElements(string $locator, ElementInterface $baseElement = null): array
     {
         $baseElement = $baseElement ?? $this->getSession()->getPage();
 
@@ -61,7 +62,7 @@ class UtilityContext extends MinkContext
             }
         );
 
-        return $elements;
+        return $elements ?? [new NullElement($locator)];
     }
 
     /**
@@ -70,11 +71,11 @@ class UtilityContext extends MinkContext
      * @param string $text Text value of the element
      * @param string $selector CSS selector of the element
      * @param string $textSelector Extra CSS selector for text of the element
-     * @param TraversableElement|null $baseElement
+     * @param ElementInterface $baseElement
      *
-     * @return NodeElement|null
+     * @return ElementInterface
      */
-    public function getElementByText(string $text, string $selector, string $textSelector = null, TraversableElement $baseElement = null): ?NodeElement
+    public function getElementByText(string $text, string $selector, string $textSelector = null, ElementInterface $baseElement = null): ElementInterface
     {
         $baseElement = $baseElement ?? $this->getSession()->getPage();
 
@@ -94,7 +95,7 @@ class UtilityContext extends MinkContext
             }
         }
 
-        return null;
+        return new NullElement($selector, $text);
     }
 
     /**
@@ -103,11 +104,11 @@ class UtilityContext extends MinkContext
      * @param string $text Fragment of text value of the element
      * @param string $selector CSS selector of the element
      * @param string $textSelector Extra CSS selector for text of the element
-     * @param TraversableElement|null $baseElement
+     * @param ElementInterface $baseElement
      *
-     * @return NodeElement|null
+     * @return ElementInterface
      */
-    public function getElementByTextFragment(string $textFragment, string $selector, string $textSelector = null, TraversableElement $baseElement = null): ?NodeElement
+    public function getElementByTextFragment(string $textFragment, string $selector, string $textSelector = null, ElementInterface $baseElement = null): ElementInterface
     {
         $baseElement = $baseElement ?? $this->getSession()->getPage();
 
@@ -127,7 +128,7 @@ class UtilityContext extends MinkContext
             }
         }
 
-        return null;
+        return new NullElement($selector, $textFragment);
     }
 
     /**
@@ -136,11 +137,11 @@ class UtilityContext extends MinkContext
      * @param string $text Text value of the element
      * @param string $selector CSS selector of the element
      * @param string $textSelector Extra CSS selector for text of the element
-     * @param TraversableElement|null $baseElement
+     * @param ElementInterface $baseElement
      *
      * @return int
      */
-    public function getElementPositionByText(string $text, string $selector, string $textSelector = null, TraversableElement $baseElement = null): int
+    public function getElementPositionByText(string $text, string $selector, string $textSelector = null, ElementInterface $baseElement = null): int
     {
         $baseElement = $baseElement ?? $this->getSession()->getPage();
         $counter = 0;
@@ -185,6 +186,8 @@ class UtilityContext extends MinkContext
         $start = time();
         $end = $start + $timeoutSeconds;
 
+        $lastInternalExceptionMessage = '';
+
         do {
             try {
                 $result = $callback($this);
@@ -193,12 +196,13 @@ class UtilityContext extends MinkContext
                     return $result;
                 }
             } catch (Exception $e) {
+                $lastInternalExceptionMessage = $e->getMessage();
             }
             usleep(250 * 1000);
         } while (time() < $end);
 
         if ($throwOnFailure) {
-            throw new Exception('Spin function did not return in time');
+            throw new Exception('Spin function did not return in time. Internal exception:' . $lastInternalExceptionMessage);
         }
     }
 
@@ -207,21 +211,22 @@ class UtilityContext extends MinkContext
      *
      * @param string $selector CSS selector for the element
      * @param int $timeout
-     * @param TraversableElement|null $baseElement Element from which the DOM will be searched
+     * @param ElementInterface $baseElement Element from which the DOM will be searched
      *
-     * @return NodeElement|null Searched element
+     * @return ElementInterface Searched element
      *
      * @throws ElementNotFoundException
      */
-    public function findElement(string $selector, int $timeout = 5, TraversableElement $baseElement = null): ?NodeElement
+    public function findElement(string $selector, int $timeout = 5, ElementInterface $baseElement = null): ElementInterface
     {
         $baseElement = $baseElement ?? $this->getSession()->getPage();
 
         try {
-            return $this->waitUntil($timeout,
+            $element = $this->waitUntil($timeout,
                 function () use ($selector, $baseElement) {
                     return $baseElement->find('css', $selector);
                 });
+            return $element ?? new NullElement($selector);
         } catch (Exception $e) {
             throw new ElementNotFoundException($this->getSession()->getDriver());
         }
@@ -230,9 +235,9 @@ class UtilityContext extends MinkContext
     /**
      * Filters an array of elements and returns the visible ones.
      *
-     * @param NodeElement[] $elements
+     * @param ElementInterface[] $elements
      *
-     * @return NodeElement[]
+     * @return ElementInterface[]
      */
     public function getVisibleElements(array $elements): array
     {
@@ -252,8 +257,10 @@ class UtilityContext extends MinkContext
         try {
             $element = $this->findElement($locator, $timeout);
 
-            return isset($element) && $element->isVisible();
+            return $element->isVisible();
         } catch (ElementNotFoundException $e) {
+            return false;
+        } catch (NullElementException $e) {
             return false;
         }
     }
