+++ src/lib/Behat/PageObject/ContentItemPage.php
@@ -19,6 +19,8 @@ class ContentItemPage extends Page
 {
     /** @var string Name by which Page is recognised */
     public const PAGE_NAME = 'ContentItemPage';
+    public const ROOT_CONTENT_NAME = 'eZ Platform';
+    public const ROOT_CONTENT_TYPE = 'Folder';
 
     /** @var RightMenu Element representing the right menu */
     public $rightMenu;
